tddc.run_tddc and tddc.__main__ were not both needed. Everything moved to tddc.__main__
